ARQ-515 OpenShift Express Container
The Container use the OpenShift Git repository for deployment, adding and pushing files to the server.

When used in combination wih source build repositories, a special marker file is added to the repository to disable the source build during testing.

The OpenShift Express environment in the current state has no management API exposed, so the Containers support for the ServletProtocol is done by introspecing the Archive on the client side to figure out which Servlets are being deployed.

The Git push return when the Server has started, but before the deployements are deployed. The container will deploy a small application that function like a ping along side the test deployment. The application will check with the internal ModelController in JBossAS7 if the requested deployment is deployed.